requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
django-modeltranslation
django-redis
celery
redis
flower
django-ckeditor-5
djangorestframework_simplejwt
django-phonenumber-field
phonenumbers
requests
cryptography